requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
requests==2.32.3
pyyaml==6.0.2
joblib==1.4.2
tqdm==4.66.4
